# tests/contract/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "BetCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "aposta": {
            "$ref": "#/components/schemas/BetResponse"
          },
          "aviso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aviso"
          },
          "casa_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casa Id"
          }
        },
        "required": [
          "aposta",
          "casa_id"
        ],
        "title": "BetCreatedResponse",
        "type": "object"
      },
...
      "BetsPageResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BetResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "BetsPageResponse",
        "type": "object"
      },
...
      "ErrorOrValidationResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          }
        ]
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Structured errors emitted by application and cross-cutting middleware.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "error": {
            "anyOf": [
              {
                "const": "rate_limited",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "retry_after": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {
        "description": "A recursively typed JSON value; never an unconstrained Any schema.",
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ],
        "title": "JsonValue"
      },
...
      "PaginationResponse": {
        "additionalProperties": false,
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginationResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CollectionToken": {
        "description": "Token opaco da extensão; armazenado no servidor somente como HMAC.",
        "in": "header",
        "name": "X-Coleta-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for authenticated betting, cash-flow, review, upload, collection, and dashboard workflows. The checked-in OpenAPI document is the compatibility contract.",
    "title": "Bancaemdia API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/apostas": {
      "get": {
        "description": "Lista apostas do usuário com filtros temporais, dimensionais e paginação.",
        "operationId": "listar_apostas_api_v1_apostas_get",
        "parameters": [
          {
            "description": "Limite inicial inclusivo do intervalo, em ISO 8601.",
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "description": "Limite final exclusivo do intervalo, em ISO 8601.",
            "in": "query",
            "name": "ate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ate"
            }
          },
          {
            "description": "Estado de liquidação da aposta usado como filtro.",
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "description": "Identificador canônico da casa usada como filtro.",
            "in": "query",
            "name": "casa_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Casa Id"
            }
          },
          {
            "description": "Identificador canônico do tipster usado como filtro.",
            "in": "query",
            "name": "tipster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipster Id"
            }
          },
          {
            "description": "Identificador canônico do mercado usado como filtro.",
            "in": "query",
            "name": "mercado_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mercado Id"
            }
          },
          {
            "description": "Identificador canônico da competição usada como filtro.",
            "in": "query",
            "name": "competicao_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Competicao Id"
            }
          },
          {
            "description": "Canal de origem da aposta usado como filtro.",
            "in": "query",
            "name": "origem",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origem"
            }
          },
          {
            "description": "Filtra apostas pela marca de revisão grave.",
            "in": "query",
            "name": "revisao_grave",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revisao Grave"
            }
          },
          {
            "description": "Inclui apostas retiradas da apuração quando verdadeiro.",
            "in": "query",
            "name": "incluir_apagadas",
            "required": false,
            "schema": {
              "default": false,
              "title": "Incluir Apagadas",
              "type": "boolean"
            }
          },
          {
            "description": "Número da página, começando em 1.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Quantidade máxima de itens retornados na página.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetsPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar apostas"
      },
      "post": {
        "description": "Cria uma aposta manual e registra o primeiro evento em seu histórico auditável.",
        "operationId": "criar_aposta_api_v1_apostas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "default": {
                  "summary": "Aposta manual",
                  "value": {
                    "casa": "betano",
                    "data_aposta": "2026-09-22T14:30:00-03:00",
                    "descricao": "Time A vence",
                    "evento": "Time A x Time B",
                    "freebet": false,
                    "mercado_bruto": "resultado final",
                    "odd": 1.95,
                    "stake_unidades": 1.0
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ApostaManual"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bet key collision."
          },
          "413": {
            "content": {
              "text/plain": {
                "example": "Content Too Large",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The request body exceeds the server-wide streaming limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Criar aposta manual"
      }
    },
...
    "/api/v1/coleta": {
      "post": {
        "description": "Alias versionado para receber um lote bruto e agendar sua materialização idempotente.",
        "operationId": "receber_coleta_api_v1_coleta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "default": {
                  "summary": "Lote da extensão",
                  "value": {
                    "apostas": [
                      {
                        "id": "BET-123",
                        "odd": 1.95,
                        "status": "open"
                      }
                    ],
                    "capturado_em": "2026-09-22T14:30:00-03:00",
                    "casa": "betano",
                    "contrato": 1
                  }
                }
              },
              "schema": {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "properties": {
                  "apostas": {
                    "items": {
                      "$ref": "#/components/schemas/JsonValue"
                    },
                    "maxItems": 1000,
                    "type": "array"
                  },
                  "capturado_em": {
                    "$ref": "#/components/schemas/JsonValue",
                    "description": "Metadado informativo enviado pela extensão; o servidor o ignora para ordenação e aceita clientes legados."
                  },
                  "casa": {
                    "maxLength": 40,
                    "minLength": 2,
                    "type": "string"
                  },
                  "contrato": {
                    "const": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "contrato",
                  "casa",
                  "apostas"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed collection payload."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid collection token."
          },
          "413": {
            "content": {
              "text/plain": {
                "example": "Content Too Large",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The request body exceeds the server-wide streaming limit."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The collection rate or daily limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The collection was saved but could not be queued."
          }
        },
        "security": [
          {
            "CollectionToken": []
          }
        ],
        "summary": "Receber coleta da extensão"
      }
    },
...
    "/coleta": {
      "post": {
        "description": "Recebe um lote bruto capturado pela extensão e agenda a materialização idempotente.",
        "operationId": "receber_coleta_coleta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "default": {
                  "summary": "Lote da extensão",
                  "value": {
                    "apostas": [
                      {
                        "id": "BET-123",
                        "odd": 1.95,
                        "status": "open"
                      }
                    ],
                    "capturado_em": "2026-09-22T14:30:00-03:00",
                    "casa": "betano",
                    "contrato": 1
                  }
                }
              },
              "schema": {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "properties": {
                  "apostas": {
                    "items": {
                      "$ref": "#/components/schemas/JsonValue"
                    },
                    "maxItems": 1000,
                    "type": "array"
                  },
                  "capturado_em": {
                    "$ref": "#/components/schemas/JsonValue",
                    "description": "Metadado informativo enviado pela extensão; o servidor o ignora para ordenação e aceita clientes legados."
                  },
                  "casa": {
                    "maxLength": 40,
                    "minLength": 2,
                    "type": "string"
                  },
                  "contrato": {
                    "const": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "contrato",
                  "casa",
                  "apostas"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed collection payload."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid collection token."
          },
          "413": {
            "content": {
              "text/plain": {
                "example": "Content Too Large",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The request body exceeds the server-wide streaming limit."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The collection rate or daily limit was exceeded."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The collection was saved but could not be queued."
          }
        },
        "security": [
          {
            "CollectionToken": []
          }
        ],
        "summary": "Receber coleta da extensão"
      }
    },
...
  }
}